// app/2026/data/types.ts
export enum RegionalStatus {
  preparing = "preparing",
  preliminariesFinished = "preliminariesFinished",
  regionalsFinished = "regionalsFinished",
}

export interface RegionalBase {
  contestDate: `2025-${string}-${string}`;
  site: string;
  region: string;
  url?: string;
  score: number;
  scoreOverride?: number;
  scoreDetails: RegionScoreArgs;
  disclaimer?: ReactNode;
  lastYear?: Regional2025;
}

export interface TeamInStandingsLike {
  teamId: number;
  rank: number | null;
  problemsSolved: number;
  institution: string;
  teamName: string;
}
// ...
export interface TeamInTeamListLikeTeamIdNullable {
  teamId: number | null;
  name: string;
  institution: string;
  problemsSolved?: number;
}

export type Regional = RegionalBase &
  (
    | {
        status: RegionalStatus.preliminariesFinished;
        regionalTeams: TeamInTeamListLikeTeamIdNullable[];
      }
    | {
        status: RegionalStatus.regionalsFinished;
        scoreboard: TeamInStandingsLike[];
      }
    | { status: RegionalStatus.preparing }
  );
// ...
export enum TeamRankInCombinedScoreboardStatus {
  NONE = "none",
  D2 = "South Pacific Independent Regional Contest",
  D3_3 = "winner universities",
  D4_2_1 = "second instance of a single team",
  D4_2_2 = "fourth or later instances of teams of a single university",
  D4_3 = "smallest value from each Asia Pacific country",
  D4_4 = "selected teams for the Championship",
  D5 = "wild-card team",
}

export interface ChampionshipTeamLike {
  teamId: number;
  assignedValue: number;
  sortKey: number;
  status: TeamRankInCombinedScoreboardStatus;
  statusMetadata?: string;
  rank: number | null;
  fromSite: string;
  rankInSite: number | null;
  rankInRegion: number | null;
  institution: string;
  teamName: string;
}